ml_project/requirements.txt
# local package
-e .

# external requirements
Sphinx>=1.0.0
coverage==6.5.0
awscli==1.27.1
python-dotenv>=0.5.1
alembic==1.8.1
asttokens==2.0.8
autopep8==2.0.0
backcall==0.2.0
boto3==1.25.2
botocore==1.28.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
docker==6.0.0
entrypoints==0.4
executing==1.1.1
flake8==5.0.4
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mlflow==1.30.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.31
protobuf==4.21.9
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
stack-data==0.5.1
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.10.0
